Fixed Full Copy not getting all records
The Full Copy had a flaw where if the MaxRecordCount was 2000 for a
Hosted Feature Service, the harvestRecords call had 1000 features hard
coded.  The problem was for every query, you missed 1000 records because
you would say "Offset by 2000, but only give me 1000".  I update the
harvestRecords method for the maxRecordCount to be passed in as a
parameter to fix this issue.